+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Examples/Consensus/Mary.hs
@@ -4,7 +4,11 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 
-module Test.Cardano.Ledger.Mary.Examples.Consensus where
+module Test.Cardano.Ledger.Api.Examples.Consensus.Mary (
+  ledgerExamplesMary,
+  exampleMultiAssetValue,
+  exampleMultiAsset,
+) where
 
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Genesis (NoGenesis (..))
@@ -14,9 +18,8 @@ import Cardano.Ledger.Mary.Value
 import qualified Data.Map.Strict as Map (singleton)
 import Data.Proxy
 import Lens.Micro
-import Test.Cardano.Ledger.Allegra.Examples.Consensus
-import Test.Cardano.Ledger.MaryEraGen ()
-import Test.Cardano.Ledger.Shelley.Examples.Consensus
+import Test.Cardano.Ledger.Api.Examples.Consensus.Allegra
+import Test.Cardano.Ledger.Api.Examples.Consensus.Shelley
 
 -- | ShelleyLedgerExamples for Allegra era
 ledgerExamplesMary :: ShelleyLedgerExamples MaryEra
@@ -37,6 +40,5 @@ exampleMultiAsset x =
   MultiAsset (Map.singleton policyId $ Map.singleton couttsCoin 1000)
   where
     policyId = PolicyID $ mkScriptHash x
-
     couttsCoin :: AssetName
     couttsCoin = AssetName "couttsCoin"