+++ eras/alonzo/test-suite/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-alonzo-test`
 
+## 1.4.0.0
+
+* Removed `Test.Cardano.Ledger.Alonzo.Examples.Consensus` (moved to `cardano-ledger-api` testlib)
+
 ## 1.3.1.1
 
 *