+++ eras/shelley-ma/test-suite/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Version history for `cardano-ledger-shelley-ma-test`
 
+## 1.4.0.0
+
+* Removed `Test.Cardano.Ledger.Allegra.Examples.Consensus` (moved to `cardano-ledger-api` testlib)
+* Removed `Test.Cardano.Ledger.Mary.Examples.Consensus` (moved to `cardano-ledger-api` testlib)
+
 ## 1.3.0.1
 
 *