+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### `testlib`
 
+* Add Consensus example modules (moved from test-packages): `Shelley`, `Allegra`, `Mary`, `Alonzo`, `Babbage`, `Conway`, `Dijkstra`
 * Re-export all Ledger `Arbitrary` instances from `Test.Cardano.Ledger.Api.Arbitrary`
 
 ## 1.11.0.0